Move Balances.tsx from organism/ to pages/
