--- package.json
@@ -1,61 +1,0 @@
-{
-  "name": "liquidator",
-  "description": "The frontend for liquidator.",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@types/node": "12.0.7",
-    "@types/react": "^16.8.20",
-    "@types/react-dom": "16.8.4",
-    "@types/react-router-dom": "^4.3.3",
-    "@types/styled-components": "^4.1.16",
-    "moment": "^2.24.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "3.0.1",
-    "styled-components": "^4.3.1",
-    "typescript": "3.5.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "storybook": "start-storybook -p 9009 -s public # auto-generated by installing storybook",
-    "build-storybook": "build-storybook -s public",
-    "json-server": "node .json-server/server.js"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@storybook/addon-actions": "^5.1.3",
-    "@storybook/addon-links": "^5.1.3",
-    "@storybook/addons": "^5.1.3",
-    "@storybook/react": "^5.1.3",
-    "@testing-library/react": "^8.0.1",
-    "@types/jest": "^24.0.15",
-    "@types/nock": "^10.0.3",
-    "@types/storybook__react": "^4.0.2",
-    "awesome-typescript-loader": "^5.2.1",
-    "jest-dom": "^3.5.0",
-    "json-server": "^0.15.0",
-    "nock": "^10.0.6",
-    "react-docgen-typescript-loader": "^3.1.0",
-    "tslint": "^5.17.0",
-    "tslint-react": "^4.0.0"
-  }
-}
